requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=0.27.0=pypi_0
altair=5.2.0=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
bottleneck=1.3.5=py311h5bb9823_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py311hd77b12b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=py311haa95532_0
cffi=1.16.0=py311h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=py311haa95532_0
contourpy=1.2.0=py311h59b6b97_0
cryptography=41.0.7=py311h89fc84f_0
cuda-cccl=12.3.101=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-opencl-dev=12.3.101=0
cuda-profiler-api=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.11.0=pyhd3eb1b0_0
diffusers=0.26.2=pypi_0
eigen=3.4.0=h59b6b97_0
ffmpeg=4.2.2=he774522_0
filelock=3.13.1=py311haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2024.2.0=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
glib=2.69.1=h5dc1a3c_2
gmpy2=2.1.2=py311h7f96b67_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=h51c971a_3
huggingface-hub=0.20.3=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py311haa95532_0
importlib-metadata=7.0.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.3=py311haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py311h2bbff1b_0
matplotlib-base=3.8.0=py311hf62ec03_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
munkres=1.1.4=py_0
networkx=3.1=py311haa95532_0
numexpr=2.8.7=py311h1fcbade_0
numpy=1.26.3=py311hdab7c0b_0
numpy-base=1.26.3=py311hd01c5d8_0
opencv=4.6.0=py311h5d08a89_5
opencv-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
packaging=23.1=py311haa95532_0
pandas=2.1.4=py311hf62ec03_0
pcre=8.45=hd77b12b_0
pillow=10.0.1=py311h045eedc_0
pip=23.3.1=py311haa95532_0
protobuf=4.25.2=pypi_0
psutil=5.9.0=py311h2bbff1b_0
py-cpuinfo=9.0.0=py311haa95532_0
pyarrow=15.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pyopenssl=23.2.0=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pysocks=1.7.1=py311haa95532_0
python=3.11.7=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.1.2=py3.11_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py311haa95532_0
pyyaml=6.0.1=py311h2bbff1b_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
referencing=0.33.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=py311haa95532_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
safetensors=0.4.2=pypi_0
scipy=1.11.4=py311hc1ccb85_0
seaborn=0.12.2=py311haa95532_0
setuptools=68.2.2=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sqlite=3.41.2=h2bbff1b_0
streamlit=1.31.0=pypi_0
sympy=1.12=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.15.1=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
torchaudio=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=pypi_0
tqdm=4.65.0=py311h746a85d_0
transformers=4.37.2=pypi_0
typing_extensions=4.9.0=py311haa95532_1
tzdata=2023d=h04d1e81_0
tzlocal=5.2=pypi_0
ultralytics=8.1.10=pypi_0
urllib3=1.26.18=py311haa95532_0
validators=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=4.0.0=pypi_0
wheel=0.41.2=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
